Move `io::IoSlice[Mut]` to `alloc`
